fix: example, features
