+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["sdk"] #"examples/*"]
+members = ["sdk", "examples/*"]
 
 
 [workspace.dependencies]
