+++ examples/system_theme/public/favicon.ico
