+++ examples/window_size/README.md
