+++ examples/window_size/public/favicon.ico
