+++ examples/window_size/src/main.rs
